moved get_null_mapping to interval.py
