+++ tests/test_interval.py
@@ -29,3 +29,11 @@ from syndiffix.interval import *
 )
 def test_snapping(raw: Interval, snapped: Interval) -> None:
     assert snap_interval(raw) == snapped
+
+def test_prepares_null_mappings() -> None:
+    assert get_null_mapping(Interval(1.2, 1.3)) == 2.6
+    assert get_null_mapping(Interval(1.2, 1.2)) == 2.4
+    assert get_null_mapping(Interval(-1.0, 3.0)) == 6.0
+    assert get_null_mapping(Interval(-4.0, -2.5)) == -8.0
+    assert get_null_mapping(Interval(-4.0, -0.0)) == -8.0
+    assert get_null_mapping(Interval(0.0, 0.0)) == 1.0
